requirements.txt
pytorch==2.1.1
pytest==7.4.3
datasets==2.12.0
numpy==1.26.2